A progress dialog now appears when setting terrain/population vertex resolution
Unfortunately, there seems to be a JavaFX glitch when running a thread indirectly triggered from a spinner. It causes the thread to be endless re-executed.

I replaced the spinners with 2 buttons and a textfield. The glitch seems to be gone.

A few tiny fixes and improvements were also made in the Population class.