Fixed issue of camera being zoomed in or out too far in a rendered image as compared to the preview
The zoom of the rendered image will not exactly match the zoom of the preview. It is approximates.

This solution should be improved. A formula could be created to determine the zoom instead of an array of constants. I have yet to successfully create a formula to do so, but I intend to eventually return to this to do so.